requirements.txt
pygame==2.5.2
openai==1.12.0
python-dotenv==1.0.0
matplotlib